backend/requirements.txt
fastapi==0.115.6
uvicorn==0.35.0
python-dotenv==1.1.0
supabase==2.20.0
chromadb==1.1.0
google-generativeai==0.8.5
pydantic==2.11.9
pdfplumber==0.11.7